event-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
kafka-python==2.0.2
pydantic==2.9.2
prometheus-fastapi-instrumentator==7.0.0
psycopg2-binary==2.9.9